Review comment changes
