--- test/utils/transformers/HouseConsignmentAdditionalReferenceTransformerSpec.scala
@@ -1,74 +1,0 @@
-/*
- * Copyright 2024 HM Revenue & Customs
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package utils.transformers
-
-import base.{AppWithDefaultMockFixtures, SpecBase}
-import connectors.ReferenceDataConnector
-import generated.AdditionalReferenceType02
-import generators.Generators
-import models.Index
-import models.reference.AdditionalReferenceType
-import org.mockito.ArgumentMatchers.{any, eq => eqTo}
-import org.mockito.Mockito.{reset, when}
-import org.scalacheck.Arbitrary.arbitrary
-import org.scalatestplus.scalacheck.ScalaCheckPropertyChecks
-import pages.houseConsignment.index.items.additionalReference.AdditionalReferencePage
-import play.api.inject.bind
-import play.api.inject.guice.GuiceApplicationBuilder
-
-import scala.concurrent.Future
-
-class HouseConsignmentAdditionalReferenceTransformerSpec extends SpecBase with AppWithDefaultMockFixtures with ScalaCheckPropertyChecks with Generators {
-
-  private val transformer: HouseConsignmentAdditionalReferenceTransformer = app.injector.instanceOf[HouseConsignmentAdditionalReferenceTransformer]
-
-  private lazy val mockRefDataConnector: ReferenceDataConnector = mock[ReferenceDataConnector]
-
-  override def guiceApplicationBuilder(): GuiceApplicationBuilder =
-    super
-      .guiceApplicationBuilder()
-      .overrides(
-        bind[ReferenceDataConnector].toInstance(mockRefDataConnector)
-      )
-
-  override def beforeEach(): Unit = {
-    super.beforeEach()
-    reset(mockRefDataConnector)
-  }
-
-  "must transform data" in {
-    val additionalReferenceType02: Seq[AdditionalReferenceType02] = arbitrary[Seq[AdditionalReferenceType02]].sample.value
-
-    additionalReferenceType02.map {
-      type0 =>
-        when(mockRefDataConnector.getAdditionalReferenceType(eqTo(type0.typeValue))(any(), any()))
-          .thenReturn(
-            Future.successful(AdditionalReferenceType(documentType = type0.typeValue, description = "describe me"))
-          )
-    }
-
-    val result = transformer.transform(additionalReferenceType02, hcIndex, itemIndex).apply(emptyUserAnswers).futureValue
-
-    additionalReferenceType02.zipWithIndex.map {
-      case (refType, i) =>
-        result.getValue(AdditionalReferencePage(hcIndex, itemIndex, Index(i))).documentType mustBe refType.typeValue
-
-    }
-
-  }
-
-}
